extract sources to separate package
