+++ docs/quickstart/quickstart2_rag.md
@@ -44,7 +44,7 @@ We first need to direct Ragbits to the location of the documents to load them. T
 
 ```python
 from pathlib import Path
-from ragbits.document_search.documents.local_file_source import LocalFileSource
+from ragbits.document_search.documents.sources.local_file_source import LocalFileSource
 
 # Path to the directory with markdown files to ingest
 documents_path = Path(__file__).parent / "pb-source/en"
