+++ packages/ragbits-document-search/src/ragbits/document_search/documents/sources/__init__.py
@@ -1,0 +1,5 @@
+from ragbits.document_search.documents.sources.gcs_source import GCSSource
+from ragbits.document_search.documents.sources.hugging_face_source import HuggingFaceSource
+from ragbits.document_search.documents.sources.local_file_source import LocalFileSource
+
+__all__ = ["GCSSource", "HuggingFaceSource", "LocalFileSource"]
