+++ packages/ragbits-document-search/src/ragbits/document_search/documents/sources/local_file_source.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from typing import ClassVar
 
 from ragbits.document_search.documents.exceptions import SourceNotFoundError
-from ragbits.document_search.documents.sources import Source
+from ragbits.document_search.documents.sources.sources import Source
 
 
 class LocalFileSource(Source):
