+++ packages/ragbits-document-search/src/ragbits/document_search/ingestion/processor_strategies/base.py
@@ -5,7 +5,7 @@ from typing import ClassVar
 from ragbits.core.utils.config_handling import WithConstructionConfig
 from ragbits.document_search.documents.document import Document, DocumentMeta
 from ragbits.document_search.documents.element import Element
-from ragbits.document_search.documents.sources import Source
+from ragbits.document_search.documents.sources.sources import Source
 from ragbits.document_search.ingestion import processor_strategies
 from ragbits.document_search.ingestion.document_processor import DocumentProcessorRouter
 from ragbits.document_search.ingestion.providers.base import BaseProvider
