+++ packages/ragbits-document-search/src/ragbits/document_search/ingestion/processor_strategies/sequential.py
@@ -2,7 +2,7 @@ from collections.abc import Sequence
 
 from ragbits.document_search.documents.document import Document, DocumentMeta
 from ragbits.document_search.documents.element import Element
-from ragbits.document_search.documents.sources import Source
+from ragbits.document_search.documents.sources.sources import Source
 from ragbits.document_search.ingestion.document_processor import DocumentProcessorRouter
 from ragbits.document_search.ingestion.providers.base import BaseProvider
