+++ packages/ragbits-document-search/tests/unit/test_sources.py
@@ -6,9 +6,9 @@ from unittest.mock import MagicMock, patch
 
 from aiohttp import ClientSession
 
-from ragbits.document_search.documents.gcs_source import GCSSource
-from ragbits.document_search.documents.hugging_face_source import HuggingFaceSource
-from ragbits.document_search.documents.sources import LOCAL_STORAGE_DIR_ENV
+from ragbits.document_search.documents.sources.gcs_source import GCSSource
+from ragbits.document_search.documents.sources.hugging_face_source import HuggingFaceSource
+from ragbits.document_search.documents.sources.sources import LOCAL_STORAGE_DIR_ENV
 
 os.environ[LOCAL_STORAGE_DIR_ENV] = Path(__file__).parent.as_posix()
 
@@ -94,7 +94,7 @@ async def test_huggingface_source_fetch() -> None:
     data = MagicMock(skip=skip)
     source = HuggingFaceSource(path="org/docs", split="train", row=1)
 
-    with patch("ragbits.document_search.documents.hugging_face_source.load_dataset", return_value=data):
+    with patch("ragbits.document_search.documents.sources.hugging_face_source.load_dataset", return_value=data):
         path = await source.fetch()
 
     assert source.id == "huggingface:org/docs/train/1"
